feat: Add smv_replacement_refs_suffix
Add `smv_refs_override_suffix` to your `conf.py`
with a value like `-docs` and if a branch or tag would
normally build, then SMV checks for a ref like
`/refs/heads/{the-good-ref}-docs` and if the ref exists,
SMV gets the archive of the repo from the replacement ref
when it builds the documentation.

Add INFO level logging so that it is clearer in the logging
when an override is used.

After troubleshooting, check for overrides using the pattern
`foo-docs` rather than `refs/heads/foo-docs`.